requirements.txt
arabic_reshaper
python_bidi